+++ PassepartoutUITests/PassepartoutUITests.swift
@@ -23,7 +23,7 @@
 //  along with Passepartout.  If not, see <http://www.gnu.org/licenses/>.
 //
 
-import UIAccessibility
+import UITesting
 import XCTest
 
 @MainActor
